Updates to the Player and Enviorment.
there is now an enviorment manager that can be used to link in biome volumes (to come) that can be used to change the temperature around the player. this adds in another layer of survival elements in which we can play with for the player. also changed instant sound wave reaction to a propagated animation wave that uses the sound waves inception point plus some physical magic to ensure that the sound wave travels at the appropriate speed. this is what actually spurred the enviorment manager because i needed a way to dynamically send temperature data to the sound wave manager. also have worked in a way to start adding in extra layers to the enviorments. for example. when rainy we can change a property gamma to another small number to effectively make the sound propagation faster or slower. as physically described sound waves are dependent on the medium that they are in. thus i have added a molarMass property too which can help with differentiating if the water is traveling through an air molecule or a water or even the walls.